Bundle fields together with a field collection
