+++ src/Plugin/resource/DataProvider/DataProviderEntity.php
@@ -14,6 +14,7 @@ use Drupal\restful\Http\HttpHeader;
 use Drupal\restful\Http\RequestInterface;
 use Drupal\restful\Plugin\resource\DataInterpreter\DataInterpreterEMW;
 use Drupal\restful\Plugin\resource\DataInterpreter\DataInterpreterInterface;
+use Drupal\restful\Plugin\resource\Field\ResourceFieldCollection;
 use Drupal\restful\Plugin\resource\Field\ResourceFieldCollectionInterface;
 use Drupal\restful\Plugin\resource\Field\ResourceFieldEntity;
 use Drupal\restful\Plugin\resource\Field\ResourceFieldEntityInterface;
@@ -219,35 +220,35 @@ class DataProviderEntity extends DataProvider implements DataProviderEntityInter
       return NULL;
     }
 
+    $resource_field_collection = new ResourceFieldCollection();
     /* @var \EntityDrupalWrapper $wrapper */
     $wrapper = entity_metadata_wrapper($this->entityType, $entity_id);
     $wrapper->language($this->getLangCode());
     $interpreter = new DataInterpreterEMW($this->getAccount(), $wrapper);
-    $values = array();
+    $resource_field_collection->setInterpreter($interpreter);
 
     $input = $this->getRequest()->getParsedInput();
     $limit_fields = !empty($input['fields']) ? explode(',', $input['fields']) : array();
 
     foreach ($this->fieldDefinitions as $resource_field_name => $resource_field) {
+      // Create an empty field collection and populate it with the appropriate
+      // resource fields.
       /* @var ResourceFieldEntityInterface $resource_field */
-      $resource_field->setInterpreter($interpreter);
 
       if ($limit_fields && !in_array($resource_field_name, $limit_fields)) {
         // Limit fields doesn't include this property.
         continue;
       }
 
-      $value = NULL;
-
       if (!$this->methodAccess($resource_field) || !$resource_field->access('view', $interpreter)) {
         // The field does not apply to the current method or has denied access.
         continue;
       }
 
-      $values[$resource_field_name] = $resource_field;
+      $resource_field_collection->set($resource_field->id(), $resource_field);
     }
 
-    return $values;
+    return $resource_field_collection;
   }
 
   /**
@@ -765,7 +766,6 @@ class DataProviderEntity extends DataProvider implements DataProviderEntityInter
 
     foreach ($this->fieldDefinitions as $public_field_name => $resource_field) {
       /* @var ResourceFieldEntityInterface $resource_field */
-      $resource_field->setInterpreter($interpreter);
 
       if (!$this->methodAccess($resource_field)) {
         // Allow passing the value in the request.
