+++ src/Plugin/resource/DataProvider/DataProviderPlug.php
@@ -71,7 +71,6 @@ class DataProviderPlug extends DataProvider implements DataProviderInterface {
 
     foreach ($this->fieldDefinitions as $resource_field_name => $resource_field) {
       /* @var ResourceFieldEntityInterface $resource_field */
-      $resource_field->setInterpreter($interpreter);
 
       if ($limit_fields && !in_array($resource_field_name, $limit_fields)) {
         // Limit fields doesn't include this property.
