+++ src/Plugin/resource/Field/ResourceFieldEntityFile.php
@@ -10,13 +10,6 @@ namespace Drupal\restful\Plugin\resource\Field;
 class ResourceFieldEntityFile extends ResourceFieldEntity implements ResourceFieldEntityInterface {
 
   /**
-   * Interpreter to use to interact with the field.
-   *
-   * @var DataInterpreterInterface
-   */
-  protected $interpreter;
-
-  /**
    * {@inheritdoc}
    */
   public function preprocess($value) {
@@ -38,22 +31,6 @@ class ResourceFieldEntityFile extends ResourceFieldEntity implements ResourceFie
       );
     }
     return $return;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function setInterpreter($interpreter) {
-    // Don't use a decorator for this, it leads to the same interpreter being
-    // assigned to the same memory object all of the results in a list call.
-    $this->interpreter = $interpreter;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getInterpreter() {
-    return $this->interpreter;
   }
 
   /**
