+++ src/Plugin/resource/Field/ResourceFieldEntityReference.php
@@ -177,8 +177,7 @@ class ResourceFieldEntityReference extends ResourceFieldEntity implements Resour
   /**
    * {@inheritdoc}
    */
-  public function value(DataInterpreterInterface $interpreter = NULL) {
-    $interpreter = $interpreter ?: $this->getInterpreter();
+  public function value(DataInterpreterInterface $interpreter) {
     $value = $this->decorated->value($interpreter);
     if (isset($value)) {
       // Let the decorated resolve callbacks.
