+++ src/Plugin/resource/Field/ResourceFieldInterface.php
@@ -129,7 +129,7 @@ interface ResourceFieldInterface {
    *
    * @throws IncompatibleFieldDefinitionException
    */
-  public function value(DataInterpreterInterface $interpreter = NULL);
+  public function value(DataInterpreterInterface $interpreter);
 
   /**
    * Check access on property by the defined access callbacks.
@@ -202,22 +202,6 @@ interface ResourceFieldInterface {
    * @throws IncompatibleFieldDefinitionException
    */
   public function set($value, DataInterpreterInterface $interpreter);
-
-  /**
-   * Set the data interpreter value for later use.
-   *
-   * @param DataInterpreterInterface $interpreter
-   *   The interpreter to use.
-   */
-  public function setInterpreter($interpreter);
-
-  /**
-   * Gets the data interpreter.
-   *
-   * @return DataInterpreterInterface
-   *   The stored interpreter.
-   */
-  public function getInterpreter();
 
   /**
    * Executes the process callbacks.
