+++ src/Plugin/resource/Field/ResourceFieldKeyValue.php
@@ -24,8 +24,7 @@ class ResourceFieldKeyValue extends ResourceField implements ResourceFieldInterf
   /**
    * {@inheritdoc}
    */
-  public function value(DataInterpreterInterface $interpreter = NULL) {
-    $interpreter = $interpreter ?: $this->getInterpreter();
+  public function value(DataInterpreterInterface $interpreter) {
     if ($value = parent::value($interpreter)) {
       return $value;
     }
